+++ src/vmm/src/cpuid/transformer/intel.rs
@@ -4,10 +4,7 @@
 use super::*;
 use crate::cpuid::bit_helper::BitHelper;
 use crate::cpuid::cpu_leaf::*;
-
-// The APIC ID shift in leaf 0xBh specifies the number of bits to shit the x2APIC ID to get a
-// unique topology of the next level. This allows 128 logical processors/package.
-const LEAFBH_INDEX1_APICID: u32 = 7;
+use crate::cpuid::transformer::common::use_host_cpuid_function;
 
 fn update_deterministic_cache_entry(
     entry: &mut kvm_cpuid_entry2,
@@ -50,95 +47,17 @@ fn update_perf_mon_entry(entry: &mut kvm_cpuid_entry2, _vm_spec: &VmSpec) -> Res
     Ok(())
 }
 
-fn update_extended_topology_entry(
-    entry: &mut kvm_cpuid_entry2,
-    vm_spec: &VmSpec,
-) -> Result<(), Error> {
-    use crate::cpuid::cpu_leaf::leaf_0xb::*;
-
-    // reset eax, ebx, ecx
-    entry.eax = 0_u32;
-    entry.ebx = 0_u32;
-    entry.ecx = 0_u32;
-    // EDX bits 31..0 contain x2APIC ID of current logical processor
-    // x2APIC increases the size of the APIC ID from 8 bits to 32 bits
-    entry.edx = u32::from(vm_spec.cpu_index);
-
-    // "If SMT is not present in a processor implementation but CPUID leaf 0BH is supported,
-    // CPUID.EAX=0BH, ECX=0 will return EAX = 0, EBX = 1 and level type = 1.
-    // Number of logical processors at the core level is reported at level type = 2."
-    // (Intel® 64 Architecture x2APIC Specification, Ch. 2.8)
-    match entry.index {
-        // Thread Level Topology; index = 0
-        0 => {
-            // To get the next level APIC ID, shift right with at most 1 because we have
-            // maximum 2 hyperthreads per core that can be represented by 1 bit.
-            entry
-                .eax
-                .write_bits_in_range(&eax::APICID_BITRANGE, u32::from(vm_spec.cpu_bits));
-            // When cpu_count == 1 or HT is disabled, there is 1 logical core at this level
-            // Otherwise there are 2
-            entry.ebx.write_bits_in_range(
-                &ebx::NUM_LOGICAL_PROCESSORS_BITRANGE,
-                u32::from(vm_spec.cpus_per_core()),
-            );
-
-            entry
-                .ecx
-                .write_bits_in_range(&ecx::LEVEL_TYPE_BITRANGE, LEVEL_TYPE_THREAD);
-        }
-        // Core Level Processor Topology; index = 1
-        1 => {
-            entry
-                .eax
-                .write_bits_in_range(&eax::APICID_BITRANGE, LEAFBH_INDEX1_APICID);
-            entry.ebx.write_bits_in_range(
-                &ebx::NUM_LOGICAL_PROCESSORS_BITRANGE,
-                u32::from(vm_spec.cpu_count),
-            );
-            entry
-                .ecx
-                .write_bits_in_range(&ecx::LEVEL_NUMBER_BITRANGE, entry.index);
-            entry
-                .ecx
-                .write_bits_in_range(&ecx::LEVEL_TYPE_BITRANGE, LEVEL_TYPE_CORE);
-        }
-        // Core Level Processor Topology; index >=2
-        // No other levels available; This should already be set correctly,
-        // and it is added here as a "re-enforcement" in case we run on
-        // different hardware
-        level => {
-            entry.ecx = level;
-        }
-    }
-
-    Ok(())
-}
-
 pub struct IntelCpuidTransformer {}
 
 impl CpuidTransformer for IntelCpuidTransformer {
     fn process_cpuid(&self, cpuid: &mut CpuId, vm_spec: &VmSpec) -> Result<(), Error> {
         // The following commit changed the behavior of KVM_GET_SUPPORTED_CPUID to no longer
-        // include leaf 0xb / sub-leaf 1.
+        // include leaf 0xb / subleaf 1.
         // https://lore.kernel.org/all/20221027092036.2698180-1-pbonzini@redhat.com/
-        // The registers within sub-leaves of 0xB are zeroed by update_extended_topology_entry()
-        // and then modified. So the value they are set to at this point doesn't matter.
-        if !cpuid
-            .as_slice()
-            .iter()
-            .any(|entry| entry.function == 0xb && entry.index == 1)
-        {
-            cpuid
-                .push(kvm_cpuid_entry2 {
-                    function: leaf_0xb::LEAF_NUM,
-                    index: 1,
-                    flags: kvm_bindings::KVM_CPUID_FLAG_SIGNIFCANT_INDEX,
-                    ..kvm_cpuid_entry2::default()
-                })
-                .map_err(Error::Fam)?;
-        }
-
+        // We call `use_host_cpuid_function()` to add the leaf 0xb / subleaf 1. As the registers
+        // within subleaves are filled by `update_extended_topology_entry()`, these values set here
+        // don't matter at this point.
+        use_host_cpuid_function(cpuid, leaf_0xb::LEAF_NUM, true)?;
         self.process_entries(cpuid, vm_spec)
     }
 
@@ -148,7 +67,7 @@ impl CpuidTransformer for IntelCpuidTransformer {
             leaf_0x4::LEAF_NUM => Some(intel::update_deterministic_cache_entry),
             leaf_0x6::LEAF_NUM => Some(intel::update_power_management_entry),
             leaf_0xa::LEAF_NUM => Some(intel::update_perf_mon_entry),
-            leaf_0xb::LEAF_NUM => Some(intel::update_extended_topology_entry),
+            leaf_0xb::LEAF_NUM => Some(common::update_extended_topology_entry),
             0x8000_0002..=0x8000_0004 => Some(common::update_brand_string_entry),
             _ => None,
         }
@@ -160,7 +79,6 @@ mod tests {
     use kvm_bindings::kvm_cpuid_entry2;
 
     use super::*;
-    use crate::cpuid::cpu_leaf::leaf_0xb::{LEVEL_TYPE_CORE, LEVEL_TYPE_THREAD};
     use crate::cpuid::transformer::VmSpec;
 
     #[test]
@@ -231,41 +149,6 @@ mod tests {
         );
     }
 
-    fn check_update_extended_topology_entry(
-        cpu_count: u8,
-        smt: bool,
-        index: u32,
-        expected_apicid: u32,
-        expected_num_logical_processors: u32,
-        expected_level_type: u32,
-    ) {
-        use crate::cpuid::cpu_leaf::leaf_0xb::*;
-
-        let vm_spec = VmSpec::new(0, cpu_count, smt).expect("Error creating vm_spec");
-        let entry = &mut kvm_cpuid_entry2 {
-            function: 0x0,
-            index,
-            flags: 0,
-            eax: 0,
-            ebx: 0,
-            ecx: 0,
-            edx: 0,
-            padding: [0, 0, 0],
-        };
-
-        assert!(update_extended_topology_entry(entry, &vm_spec).is_ok());
-
-        assert!(entry.eax.read_bits_in_range(&eax::APICID_BITRANGE) == expected_apicid);
-        assert!(
-            entry
-                .ebx
-                .read_bits_in_range(&ebx::NUM_LOGICAL_PROCESSORS_BITRANGE)
-                == expected_num_logical_processors
-        );
-        assert!(entry.ecx.read_bits_in_range(&ecx::LEVEL_TYPE_BITRANGE) == expected_level_type);
-        assert!(entry.ecx.read_bits_in_range(&ecx::LEVEL_NUMBER_BITRANGE) == index);
-    }
-
     #[test]
     fn test_1vcpu_ht_off() {
         // test update_deterministic_cache_entry
@@ -275,12 +158,6 @@ mod tests {
         check_update_deterministic_cache_entry(1, false, 2, 0);
         // test L3
         check_update_deterministic_cache_entry(1, false, 3, 0);
-
-        // test update_extended_topology_entry
-        // index 0
-        check_update_extended_topology_entry(1, false, 0, 0, 1, LEVEL_TYPE_THREAD);
-        // index 1
-        check_update_extended_topology_entry(1, false, 1, LEAFBH_INDEX1_APICID, 1, LEVEL_TYPE_CORE);
     }
 
     #[test]
@@ -292,12 +169,6 @@ mod tests {
         check_update_deterministic_cache_entry(1, true, 2, 0);
         // test L3
         check_update_deterministic_cache_entry(1, true, 3, 0);
-
-        // test update_extended_topology_entry
-        // index 0
-        check_update_extended_topology_entry(1, true, 0, 0, 1, LEVEL_TYPE_THREAD);
-        // index 1
-        check_update_extended_topology_entry(1, true, 1, LEAFBH_INDEX1_APICID, 1, LEVEL_TYPE_CORE);
     }
 
     #[test]
@@ -309,12 +180,6 @@ mod tests {
         check_update_deterministic_cache_entry(2, false, 2, 1);
         // test L3
         check_update_deterministic_cache_entry(2, false, 3, 1);
-
-        // test update_extended_topology_entry
-        // index 0
-        check_update_extended_topology_entry(2, false, 0, 0, 1, LEVEL_TYPE_THREAD);
-        // index 1
-        check_update_extended_topology_entry(2, false, 1, LEAFBH_INDEX1_APICID, 2, LEVEL_TYPE_CORE);
     }
 
     #[test]
@@ -326,11 +191,5 @@ mod tests {
         check_update_deterministic_cache_entry(2, true, 2, 0);
         // test L3
         check_update_deterministic_cache_entry(2, true, 3, 0);
-
-        // test update_extended_topology_entry
-        // index 0
-        check_update_extended_topology_entry(2, true, 0, 1, 2, LEVEL_TYPE_THREAD);
-        // index 1
-        check_update_extended_topology_entry(2, true, 1, LEAFBH_INDEX1_APICID, 2, LEVEL_TYPE_CORE);
     }
 }
